refactored other commands
